refactor: combine all guides
* move welcome guide into school guide
* remove references into page/index.tsx
* split all links into its own file
