+++ src/components/Modal/Home/SchoolGuide/components/index.tsx
@@ -3,8 +3,11 @@ type TABS = {
 };
 
 export { default as SchoolSurvival } from './SchoolSurvival';
+export { default as SoftwareGuide } from './SoftwareGuide';
+export { default as WelcomeGuide } from './WelcomeGuide';
 
 export const AVAILABLE_TABS: TABS = {
   daily: 'Daily',
+  welcome: 'Newcomers',
   software: 'Software',
 };
