--- src/utils/guidesLinks.tsx
@@ -1,54 +1,0 @@
-export const School_Guide_Links = [
-  {
-    id: 'apa_citation_web',
-    title: 'APA Citation',
-    logo: 'public',
-    type: 'url',
-    link: 'https://researchguides.smu.edu.sg/c.php?g=421741&p=6627438',
-  },
-  {
-    id: 'academic_handbook_web',
-    title: 'Student Academic Handbook',
-    logo: 'public',
-    type: 'url',
-    link: 'https://inet.smu.edu.sg/sites/courses/Pages/Student-Academic-Handbook.aspx',
-  },
-  {
-    id: 'boss_guide_web',
-    title: 'BOSS Bidding Guidebook 2020',
-    logo: 'public',
-    type: 'url',
-    link: 'http://bit.ly/smusa-bossbidding-guidebook2',
-  },
-  {
-    id: 'reset_password_web',
-    title: 'Password Reset Service',
-    logo: 'public',
-    type: 'url',
-    link: 'https://pwdreset.smu.edu.sg/reset/reset',
-  },
-];
-
-export const Welcome_Guide_Links = [
-  {
-    id: 'concession_card_web',
-    title: 'Concession Card',
-    logo: 'commute',
-    type: 'url',
-    link: 'https://matriculation.smu.edu.sg/2021/tertiary-student-ez-link-card',
-  },
-  {
-    id: 'eduroam_web',
-    title: 'Eduroam',
-    logo: 'public',
-    type: 'url',
-    link: 'https://itsupport.smu.edu.sg/hc/en-us/articles/203357860-What-is-eduroam-',
-  },
-  {
-    id: 'internet_access_web',
-    title: 'Internet Access',
-    logo: 'public',
-    type: 'url',
-    link: 'https://itsupport.smu.edu.sg/hc/en-us/articles/204476054-How-can-I-connect-to-SMU-Wireless-networks-',
-  },
-];
