+++ src/utils/schoolGuide/dailyLinks.tsx
@@ -1,0 +1,27 @@
+export const School_Guide_Links = [
+  {
+    id: 'apa_citation_web',
+    title: 'APA Citation',
+    link: 'https://researchguides.smu.edu.sg/c.php?g=421741&p=6627438',
+  },
+  {
+    id: 'reset_password_web',
+    title: 'Password Reset Service',
+    link: 'https://pwdreset.smu.edu.sg/reset/reset',
+  },
+  {
+    id: 'rent_a_locker',
+    title: 'Rent a Locker',
+    link: 'https://rent-a-locker.com/',
+  },
+  {
+    id: 'enable-2fa',
+    title: '2FA',
+    link: 'https://register2fa.smu.edu.sg/MultiFactorAuth/',
+  },
+  {
+    id: 'student_withdrawal_form',
+    title: 'Student Withdrawal Form',
+    link: 'https://oasis.smu.edu.sg/Documents/Downloads/RO/Withdrawal%20Form.pdf',
+  },
+];
