+++ src/utils/schoolGuide/welcomeLinks.tsx
@@ -1,0 +1,32 @@
+export const Welcome_Guide_Links = [
+  {
+    id: 'concession_card_web',
+    title: 'Concession Card',
+    link: 'https://matriculation.smu.edu.sg/2021/tertiary-student-ez-link-card',
+  },
+  {
+    id: 'eduroam_web',
+    title: 'Eduroam',
+    link: 'https://itsupport.smu.edu.sg/hc/en-us/articles/203357860-What-is-eduroam-',
+  },
+  {
+    id: 'internet_access_web',
+    title: 'Internet Access',
+    link: 'https://itsupport.smu.edu.sg/hc/en-us/articles/204476054-How-can-I-connect-to-SMU-Wireless-networks-',
+  },
+  {
+    id: 'academic_handbook_web',
+    title: 'Student Academic Handbook',
+    link: 'https://inet.smu.edu.sg/sites/courses/Pages/Student-Academic-Handbook.aspx',
+  },
+  {
+    id: 'boss_guide_web',
+    title: 'BOSS Bidding Guidebook 2020',
+    link: 'http://bit.ly/smusa-bossbidding-guidebook2',
+  },
+  {
+    id: 'housing_application',
+    title: 'Housing Application',
+    link: 'https://housing.intranet.smu.edu.sg/apply/',
+  },
+];
